+++ scripts/export.lua
@@ -7,19 +7,19 @@ local i18n = require 'lib.i18n'
 local fun = require 'lib.fun'
 require 'lib.table'
 
+local Export = {}
+
 local function create_zip(fp)
   local zipfile, err = Zip.new(fp)
   Logs.assert(zipfile, i18n 'export.zip_create_error', ' ', err)
   return zipfile
 end
 
-local function get_outdir(flag_o)
-  return flag_o and flag_o[1] or '.'
-end
+local function get_outdir(flag_o) return flag_o and flag_o[1] or '.' end
 
 local function get_output(outdir, exp, picset)
   local zipname = ('%s-%s'):format(exp, picset)
-  return path.join(outdir, zipname .. '.zip'), zipname
+  return path.join(outdir, zipname .. '.zip')
 end
 
 local function scan_dir(dir, pattern, out)
@@ -49,40 +49,57 @@ local function scan_pics(id, picset)
   return pics
 end
 
-local function get_expansion_file(exp)
-  exp = exp .. '.cdb'
+local function get_expansion_file(eid)
+  local exp = eid .. '.cdb'
   local fp = path.join('expansions', exp)
   return fun {{fp, fp}}
 end
 
-return function(flags)
-  local fe, fp = flags['-Eall'] or flags['-e'], flags['-Pall'] or flags['-p']
-  local verbose = flags['-v'] or flags['--verbose']
-  local outdir = get_outdir(flags['-o'])
-  local picsets = Config.groups.from_flag.get_many('picset', fp)
-  local expansions = Config.groups.from_flag.get_many('expansion', fe)
+local function get_strings_file(eid)
+  local fp = path.join('expansions', eid .. '-strings.conf')
+  return fun {{fp, fp}}
+end
+
+function Export.export(outdir, expansions, picsets, verbose)
   local has_any = next(expansions) and next(picsets)
   local scripts = has_any and scan_scripts()
-  for exp, _ in pairs(expansions) do
-    local exp_file = get_expansion_file(exp)
-    for id, picset in pairs(picsets) do
-      Logs.info(i18n('export.status', {exp, id}))
-      local out, zipname = get_output(outdir, exp, id)
+  for eid, _ in pairs(expansions) do
+    local exp_file = get_expansion_file(eid)
+    local str_file = get_strings_file(eid)
+    for pid, picset in pairs(picsets) do
+      Logs.info(i18n('export.status', {eid, pid}))
+      local out = get_output(outdir, eid, pid)
       local zipfile = create_zip(out)
-      local files = fun {exp_file, scripts, scan_pics(id, picset)}
+      local files = fun {exp_file, str_file, scripts, scan_pics(pid, picset)}
                       :reduce(fun {}, fun 'a, b -> a .. b')
       local bar = Logs.bar(#files)
+      local errors = fun {}
       for _, f in ipairs(files) do
-        f = {f[1], path.join(zipname, f[2])}
-        bar:update('', verbose and i18n('export.file_srcdst', f))
+        bar:update(f[1], verbose and i18n('export.file_srcdst', f))
         local ok, err = zipfile:add(unpack(f))
-        Logs.assert(ok, i18n 'export.zip_add_error', ' ', err)
+        if not ok then errors:push(err .. '\n') end
       end
       bar:finish()
+      if #errors > 0 then
+        local warn = i18n('export.zip_add_error', {count = #errors})
+        Logs.warning(warn, '\n', unpack(errors))
+      end
       zipfile:close()
     end
   end
-  if has_any then
+  return has_any
+end
+
+function Export.run(flags)
+  local fe, fp = flags['-Eall'] or flags['-e'], flags['-Pall'] or flags['-p']
+  local verbose = flags['--verbose']
+  local outdir = get_outdir(flags['-o'])
+  local picsets = Config.groups.from_flag.get_many('picset', fp)
+  local expansions = Config.groups.from_flag.get_many('expansion', fe)
+  if Export.export(outdir, expansions, picsets, verbose) then
     Logs.ok(i18n 'export.done')
   end
 end
+
+setmetatable(Export, {__call = Export.run})
+return Export