+++ scripts/sync.lua
@@ -1,162 +1,78 @@
 local Logs = require 'lib.logs'
 local Config = require 'scripts.config'
-local fs = require 'lib.fs'
-local path = fs.path
+local Export = require 'scripts.export'
+local path = require 'lib.fs'.path
 local i18n = require 'lib.i18n'
+local fun = require 'lib.fun'
 
-local function cp(src, dst)
-  local fsrc = io.open(src, 'rb')
-  if not fsrc then
-    return 0
-  end
-  local fdst = io.open(dst, 'wb')
-  if not fdst then
-    return 0
-  end
-  fdst:write(fsrc:read('*a'))
-  fsrc:close()
-  fdst:close()
-  return 1
-end
-
--- TODO: Refactor this into something simpler, and using the progress bar
-local function copy_dir(pattern, src, dst, tags)
-  local function cpd()
-    local to_copy = {}
-    local copied, total = 0, 0
-    for entry in fs.dir(src) do
-      if entry:match(pattern) then
-        table.insert(to_copy, entry)
-        total = total + 1
-      end
-    end
-    for _, file in ipairs(to_copy) do
-      copied = copied + cp(path.join(src, file), path.join(dst, file))
-    end
-    return copied, total
-  end
-  local s, copied, total = pcall(cpd)
-  if s then
-    Logs.info(('%d out of %d %s copied for %q'):format(copied, total, tags[1],
-                                                       tags[2]))
-    if copied == 0 then
-      Logs.warning('No ', tags[1], ' copied for this gamedir.')
-    end
-  else
-    Logs.warning(('Failed while copying %s for %q:\n'):format(tags[1], tags[2]),
-                 copied)
-  end
-end
-
-local function copy_scripts(gamedir, gpath)
-  copy_dir('c%d+%.lua', 'script', path.join(gpath, 'script'),
-           {'scripts', gamedir})
-end
-
-local function copy_pics(gamedir, gpath, picset, pscfg)
-  if not pscfg.ext then
-    pscfg.ext = 'jpg'
-  end
-  copy_dir('%d+%.' .. pscfg.ext, path.join('pics', picset),
-           path.join(gpath, 'pics'), {'pics', gamedir})
-  if pscfg.field then
-    copy_dir('%d+%.' .. pscfg.ext, path.join('pics', picset, 'field'),
-             path.join(gpath, 'pics', 'field'), {'field pics', gamedir})
-  end
-end
-
-local function copy_expansion(gamedir, gpath, exp)
-  local expansion = exp .. '.cdb'
-  local src = path.join('expansions', expansion)
-  local dst = path.join(gpath, 'expansions', expansion)
-  if cp(src, dst) == 1 then
-    Logs.info('Copied expansion for "', gamedir, '"')
-  else
-    Logs.warning('Failed to copy expansion')
-  end
-end
-
-local function get_set_codes()
-  local src = io.open(path.join('expansions', 'strings.conf'))
-  if not src then
-    return
-  end
-  local setcodes, unwritten = {}, {}
+local STRING_KEYS = fun {setname = true, counter = true}
+local function get_lines(fp)
+  local src = io.open(fp)
+  if not src then return end
+  local lines = STRING_KEYS:map(fun '_ -> {}')
+  local unwritten = STRING_KEYS:map(fun '_ -> {}')
   for line in src:lines() do
-    local code, name = line:match('^%s*!setname%s+(0x%x+)%s*(.*)$')
+    local key, code, val = line:match('^%s*!(%w+)%s+(0x%x+)%s*(.-)%s*$')
     code = tonumber(code)
-    if code and name then
-      setcodes[code], unwritten[code] = name, name
+    if STRING_KEYS[key] and code and val then
+      lines[key][code], unwritten[key][code] = val, val
     end
   end
   src:close()
-  return setcodes, unwritten
+  return lines, unwritten
 end
 
-local function get_merged_sets(fp, setcodes, unwritten)
-  local f, lines = io.open(fp), ''
+local fmt_line = fun '... -> ("!%s 0x%04x %s"):format(...)'
+local function get_merged_lines(fp, rlines, unwritten)
+  local f, wlines = io.open(fp), fun {}
   if f then
     for line in f:lines() do
-      local code = tonumber(line:match('^%s*!setname%s+(0x%x+).*$') or '')
-      local name = setcodes[code]
-      if name then
-        unwritten[code] = nil
-        lines = lines .. ('!setname 0x%04x %s\n'):format(code, name)
+      local key, code = line:match('^%s*!(%w+)%s+(0x%x+).*$')
+      code = tonumber(code or '')
+      local val = rlines[key] and rlines[key][code]
+      if val then
+        unwritten[key][code] = nil
+        wlines:push(fmt_line(key, code, val))
       else
-        lines = lines .. line .. '\n'
+        wlines:push(line)
       end
     end
     f:close()
   end
-  for code, name in pairs(unwritten) do
-    lines = lines .. ('!setname 0x%04x %s\n'):format(code, name)
+  for key, t in pairs(unwritten) do
+    for code, val in pairs(t) do
+      wlines:push(fmt_line(key, code, val))
+    end
   end
-  return lines
+  return table.concat(wlines, '\n')
 end
 
-local function copy_strings(gamedir, gpath)
-  local setcodes, unwritten = get_set_codes()
-  if not setcodes then
-    return
-  end
-  local target = path.join(gpath, 'expansions', 'strings.conf')
-  local lines = get_merged_sets(target, setcodes, unwritten)
-  local dst = io.open(target, 'w')
-  if not dst then
-    return
-  end
-  dst:write(lines)
-  -- Logs.info('Written strings.conf for "', gamedir, '"')
+local function copy_strings(eid, gid, gpath)
+  local src_fp = path.join('expansions', eid .. '-strings.conf')
+  local rlines, unwritten = get_lines(src_fp)
+  if not rlines then return end
+  local dst_fp = path.join(gpath, 'expansions', 'strings.conf')
+  local wlines = get_merged_lines(dst_fp, rlines, unwritten)
+  local dst = io.open(dst_fp, 'w')
+  if not dst then return end
+  Logs.info(i18n 'sync.writing_string')
+  dst:write(wlines)
   dst:close()
 end
 
 return function(flags)
   local fg, fp, fe = flags['-Gall'] or flags['-g'], flags['-p'], flags['-e']
-  local no_script = flags['--no-script']
-  local no_pics = flags['--no-pics']
-  local no_exp = flags['--no-exp']
   local no_string = flags['--no-string']
+  local verbose = flags['--verbose']
   local gamedirs = Config.groups.from_flag.get_many('gamedir', fg)
-  local picset, pscfg, exp
-  if not no_pics then
-    picset, pscfg = Config.groups.from_flag.get_one('picset', fp)
-  end
-  if not no_exp then
-    exp = Config.groups.from_flag.get_one('expansion', fe)
-  end
-  for gd, gdcfg in pairs(gamedirs) do
-    Logs.info(i18n('sync.status', {picset, exp, gd}))
-    if not no_script then
-      copy_scripts(gd, gdcfg.path)
-    end
-    if not no_pics then
-      copy_pics(gd, gdcfg.path, picset, pscfg)
-    end
-    if not no_exp then
-      copy_expansion(gd, gdcfg.path, exp)
-    end
+  local pid, picset = Config.groups.from_flag.get_one('picset', fp)
+  local eid, exp = Config.groups.from_flag.get_one('expansion', fe)
+  for gid, gamedir in pairs(gamedirs) do
+    Logs.info(i18n('sync.status', {pid, eid, gid}))
+    local outdir = path.join(gamedir.path, 'expansions')
+    Export.export(outdir, {[eid] = exp}, {[pid] = picset}, verbose)
     if not no_string then
-      copy_strings(gd, gdcfg.path)
+      copy_strings(eid, gid, gamedir.path)
     end
   end
   Logs.ok(i18n 'sync.done')